Update test_game.py
GitHub Repository - Python Test Game

Welcome to the GitHub repository for a Python test game program! This program challenges users to fill in the correct words and defines success as providing the correct answers. The primary objective is to test the user's vocabulary and language skills.

In order to play the game, users are presented with a series of word prompts or sentences with blanks to be filled. The player is required to input their answers within a certain time limit. If the provided word is correct, the program will proceed to the next question. However, if the user's response is incorrect, the program will indicate a failure and prompt the user to move to the next question.

Feel free to explore the repository and make any contributions or suggestions you deem necessary. Happy gaming!